--- src/main/java/com/hlxndr/jobsearchapplication/configs/SwaggerConst.java
@@ -1,17 +1,0 @@
-package com.hlxndr.jobsearchapplication.configs;
-
-public class SwaggerConst {
-
-    public static final String CONTACT_EMAIL = "email@outlook.com";
-    public static final String CONTACT_URL = "https://www.jobsearch.com";
-    public static final String CONTACT_NAME = "Job search API Support";
-    public static final String API_TITLE = "Job Searching API";
-    public static final String API_DESCRIPTION = "The application enables the registration of client-applications " +
-            "for creating job opportunities. After registration the client gets an Api-key for accessing further endpoints.\n" ;
-    public static final String TERM_OF_SERVICE = "Term of service";
-    public static final String API_VERSION = "1.0";
-    public static final String LICENSE = "Apache License 2.1.0";
-    public static final String LICENSE_URL = "https://www.apache.org/licenses/LICENSE-2.0";
-
-
-}
