+++ src/main/java/com/hlxndr/jobsearchapplication/model/ClientApp.java
@@ -23,9 +23,4 @@ public class ClientApp {
     @Size(min = 3, max = 100, message = "Name must be between 3 and 100 characters")
     private String name;
     private UUID ownApiKey;
-
-    public ClientApp(String email, String name) {
-        this.email = email;
-        this.name = name;
-    }
 }
