+++ src/main/resources/application.properties
@@ -2,8 +2,8 @@ server.port=8081
 spring.mvc.pathmatch.matching-strategy = ANT_PATH_MATCHER
 spring.datasource.url=jdbc:h2:mem:jobsearchdb
 spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=${DB_USER}
-spring.datasource.password=${DB_PASSWORD}
+spring.datasource.username=root
+spring.datasource.password=password
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 
 spring.h2.console.enabled=true
