--- src/test/java/com/hlxndr/jobsearchapplication/controller/JobSearchControllerTest.java
@@ -1,25 +1,0 @@
-package com.hlxndr.jobsearchapplication.controller;
-
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ContextConfiguration;
-
-@SpringBootTest
-@ContextConfiguration
-@AutoConfigureMockMvc
-class JobSearchControllerTest {
-
-    @Test
-    void registerClient() {
-    }
-
-    @Test
-    void addPosition() {
-    }
-
-    @Test
-    void listSearchedPositions() {
-    }
-}
