--- src/test/java/com/hlxndr/jobsearchapplication/service/JobSearchServiceTest.java
@@ -1,28 +1,0 @@
-package com.hlxndr.jobsearchapplication.service;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class JobSearchServiceTest {
-
-    @Test
-    void registerClient() {
-    }
-
-    @Test
-    void addPosition() {
-    }
-
-    @Test
-    void findPositionById() {
-    }
-
-    @Test
-    void findByNameAndLocation() {
-    }
-
-    @Test
-    void listJobUrls() {
-    }
-}
